Refactor maker editor into modular pieces
